refactor(#267 F4): extract TTS orchestrator from tts/mod.rs into tts/say.rs
The audit (#267 P1.F4): `tts/mod.rs` was 593 LOC mixing the module
façade (pub mod declarations + public types) with the synthesis
dispatcher itself (`say`, `say_ssml`, `synth_segments_*`,
`synth_one_*`, `say_with_*`, helpers). "Where does `say()` live?" →
"in `tts/mod.rs`" is the smell.

Move the dispatcher to `tts/say.rs`. `tts/mod.rs` becomes a thin
façade (~80 LOC) that:

  - Declares the per-engine submodules (`kokoro`, `vosk`, `avspeech`,
    `en`, `ru`, `g2p`, `ssml`, `sessions`, `encode`, etc).
  - Owns the public types (`TtsError`, `EngineChoice`, `SayOptions`,
    `MAX_TEXT_CHARS`).
  - Re-exports the entry points: `pub use say::{say,
    synth_segments_kokoro_with, synth_segments_vosk_with}`.

External callers (`main.rs`, `say_loop.rs`) keep reaching `tts::say()`
unchanged via the re-export. Internal helpers (`silence_samples`,
`compose_rate`, `say_ssml`, `synth_one_*`, `say_with_*`,
`synth_segments_vosk`, `encode_or_fail`, `transcode_to`,
`wav_to_mono_f32`) are private to `tts/say.rs` — they were already
private inside `mod.rs`, the move just gives them a proper home.

`MAX_BREAK_SECS` and the `compose_rate` unit test follow their fn into
`say.rs`.

Gates:
  cargo fmt --check                            clean
  cargo clippy --all-targets -- -D warnings    clean (default features)
  cargo test --lib tts                         162 / 162 passing

Refs #267
